+++ tests/integration/cdn_dedicated_waf/test_cdn_dedicated_waf_provisioning.py
@@ -49,6 +49,18 @@ from tests.lib.cdn.update import (
     subtest_update_same_domains_updates_cloudfront,
     subtest_update_same_domains_does_not_delete_server_certificate,
 )
+from tests.integration.cdn_dedicated_waf.provision import (
+    subtest_provision_create_web_acl,
+    subtest_provision_creates_health_checks,
+    subtest_provision_associates_health_checks,
+)
+from tests.integration.cdn_dedicated_waf.update import (
+    subtest_update_web_acl_does_not_update,
+    subtest_updates_health_checks,
+    subtest_updates_health_checks_do_not_change,
+    subtest_updates_associated_health_checks,
+    subtest_updates_associated_health_checks_no_change,
+)
 
 # The subtests below are "interesting".  Before test_provision_happy_path, we
 # had separate tests for each stage in the task pipeline.  But each test would
@@ -210,192 +222,3 @@ def subtest_update_same_domains(
     subtest_updates_health_checks_do_not_change(tasks, route53, instance_model)
     subtest_updates_associated_health_checks_no_change(tasks, shield, instance_model)
     subtest_update_marks_update_complete(tasks, instance_model)
-
-
-def subtest_provision_create_web_acl(tasks, wafv2):
-    db.session.expunge_all()
-    service_instance = db.session.get(CDNDedicatedWAFServiceInstance, "4321")
-
-    wafv2.expect_create_web_acl(
-        service_instance.id,
-        config.WAF_RATE_LIMIT_RULE_GROUP_ARN,
-    )
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(CDNDedicatedWAFServiceInstance, "4321")
-    assert service_instance.dedicated_waf_web_acl_id
-    assert (
-        service_instance.dedicated_waf_web_acl_id
-        == f"{service_instance.id}-dedicated-waf-id"
-    )
-    assert service_instance.dedicated_waf_web_acl_name
-    assert (
-        service_instance.dedicated_waf_web_acl_name
-        == f"{service_instance.id}-dedicated-waf"
-    )
-    assert service_instance.dedicated_waf_web_acl_arn
-    assert (
-        service_instance.dedicated_waf_web_acl_arn
-        == f"arn:aws:wafv2::000000000000:global/webacl/{service_instance.id}-dedicated-waf"
-    )
-
-
-def subtest_update_web_acl_does_not_update(tasks, wafv2):
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    # Nothing should get updated since the domains have not changed
-    wafv2.assert_no_pending_responses()
-
-
-def subtest_provision_creates_health_checks(tasks, route53, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-
-    for domain_name in service_instance.domain_names:
-        route53.expect_create_health_check(service_instance.id, domain_name)
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert sorted(
-        service_instance.route53_health_checks,
-        key=lambda check: check["domain_name"],
-    ) == [
-        {
-            "domain_name": "example.com",
-            "health_check_id": "example.com ID",
-        },
-        {"domain_name": "foo.com", "health_check_id": "foo.com ID"},
-    ]
-    route53.assert_no_pending_responses()
-
-
-def subtest_updates_health_checks(tasks, route53, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-
-    route53.expect_create_health_check(service_instance.id, "bar.com")
-
-    delete_health_check = [
-        check
-        for check in service_instance.route53_health_checks
-        if check["domain_name"] == "example.com"
-    ][0]
-    route53.expect_delete_health_check(delete_health_check["health_check_id"])
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert sorted(
-        service_instance.route53_health_checks,
-        key=lambda check: check["domain_name"],
-    ) == [
-        {
-            "domain_name": "bar.com",
-            "health_check_id": "bar.com ID",
-        },
-        {"domain_name": "foo.com", "health_check_id": "foo.com ID"},
-    ]
-
-    route53.assert_no_pending_responses()
-
-
-def subtest_updates_health_checks_do_not_change(tasks, route53, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-
-    health_checks_pre_update = service_instance.route53_health_checks
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    route53.assert_no_pending_responses()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert service_instance.route53_health_checks == health_checks_pre_update
-
-
-def subtest_provision_associates_health_checks(tasks, shield, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    if not service_instance:
-        raise Exception("Could not load service instance")
-
-    protection_id = str(uuid.uuid4())
-    protection = {
-        "Id": protection_id,
-        "ResourceArn": service_instance.cloudfront_distribution_arn,
-    }
-    shield.expect_list_protections([protection])
-
-    for health_check in service_instance.route53_health_checks:
-        health_check_id = health_check["health_check_id"]
-        shield.expect_associate_health_check(protection_id, health_check_id)
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert service_instance.shield_associated_health_checks == [
-        {
-            "health_check_id": "example.com ID",
-            "protection_id": protection_id,
-        },
-        {
-            "health_check_id": "foo.com ID",
-            "protection_id": protection_id,
-        },
-    ]
-    shield.assert_no_pending_responses()
-
-
-def subtest_updates_associated_health_checks(tasks, shield, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    if not service_instance:
-        raise Exception("Could not load service instance")
-
-    # get protection ID from initial creation
-    protection_id = service_instance.shield_associated_health_checks[0]["protection_id"]
-
-    shield.expect_associate_health_check(protection_id, "bar.com ID")
-    shield.expect_disassociate_health_check(protection_id, "example.com ID")
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert sorted(
-        service_instance.shield_associated_health_checks,
-        key=lambda check: check["health_check_id"],
-    ) == [
-        {
-            "health_check_id": "bar.com ID",
-            "protection_id": protection_id,
-        },
-        {
-            "health_check_id": "foo.com ID",
-            "protection_id": protection_id,
-        },
-    ]
-    shield.assert_no_pending_responses()
-
-
-def subtest_updates_associated_health_checks_no_change(tasks, shield, instance_model):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    if not service_instance:
-        raise Exception("Could not load service instance")
-
-    checks_pre_update = service_instance.shield_associated_health_checks
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    shield.assert_no_pending_responses()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    assert service_instance.shield_associated_health_checks == checks_pre_update
